backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
anthropic==0.42.0
faster-whisper==1.1.0
ffmpeg-python==0.2.0
pydantic-settings==2.7.0
loguru==0.7.3
python-multipart==0.0.20
alembic==1.14.0
python-dotenv==1.0.1
httpx>=0.27
psutil>=5.9
